+++ e2e/setup/index.ts
@@ -1,6 +1,9 @@
 import setupApp from './setupApp';
+import setupCodegen from './setupCodegen';
 import setupCoverage from './setupCoverage';
 import setupSelectors from './setupSelectors';
+
+export { setupApp, setupCodegen, setupCoverage, setupSelectors };
 
 export default () => {
   // call first
@@ -9,4 +12,5 @@ export default () => {
   setupCoverage();
   // call at the end - navigates the page
   setupApp();
+  !process.env.CI && setupCodegen();
 };