+++ e2e/tests/template/index.spec.ts
@@ -1,5 +1,6 @@
-import { test } from '@playwright/test';
+import { expect, test } from '@playwright/test';
 import setup from '../../setup';
+import type setupCodegen from '../../setup/setupCodegen';
 import { CanvasUtil } from '../../utils/CanvasUtil';
 import { ObjectUtil } from '../../utils/ObjectUtil';
 
@@ -11,3 +12,15 @@ test('TEST NAME', async ({ page }) => {
   const textboxUtil = new ObjectUtil(page, 'textbox');
   // write the test
 });
+
+/**
+ * ## Codegen
+ *
+ * This test supports code generation of tests
+ * see {@link setupCodegen} for uage information
+ */
+!process.env.CI &&
+  test.only('Codegen', async ({ page }) => {
+    // opens playwright codegen devtools
+    await page.pause();
+  });