+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [next]
 
+- cd(): playwright codegen [#9350](https://github.com/fabricjs/fabric.js/pull/9350)
 - tests(): migrate target hit tests to jest and drag and drop test to playwright [#9333](https://github.com/fabricjs/fabric.js/pull/9333)
 - fix(SVGParser): avoid crashing on SVG that use @import css feature [#9602](https://github.com/fabricjs/fabric.js/pull/9602)
 - fix(): compositionEnd event handler is not registered correctly. (regression from f91362c ) [#9610](https://github.com/fabricjs/fabric.js/pull/9610)